requirements.txt
# Core scheduling and automation
APScheduler==3.10.4

# Configuration management
PyYAML==6.0.1

# Git operations
GitPython==3.1.40

# Email functionality
secure-smtplib==0.1.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Add your LinkedIn bot dependencies below:
# selenium==4.15.2
# beautifulsoup4==4.12.2
# etc.
